travis: wrap a subset of builds in scan-build
The scan-build script runs clang's static analyzer on the
compiled source after each compilation.

scan-build is best done in Debug mode, because it can use assertions
like assert(some_pointer) as a hint that a pointer cannot legitimately
be null. To benefit from that, switch the 32-bit Windows build and the
clang build to Debug mode.

The 32-bit Linux and 64-bit Windows builds, together with one of the
three 64-bit Linux builds, remain in release mode so that we still
exercise that code path, and so that the build will fail sooner if
a branch doesn't compile at all (one one platform or on one
pointer-width).

Signed-off-by: Simon McVittie <smcv@debian.org>